Switching to using ParticleMass for deposition in particle_density and changing CICDeposit_3 to use 64 bit buffers everywhere.
This should ensure that ParticleMass, which is a yt-defined field that is in g,
will be used instead of the Enzo-specific particle_mass, which is actually a
density.  Additionally, because we're going to be passing around grams instead
of Msun, we use 64-bit buffers in CICDeposit_3.  This may require a change to
the particle interpolation routines from John ZuHone as well, which I will
leave up to a later changeset.

I believe I've caught all of the fields that use CICDeposit_3.

Note also that the _convertDensity function is no longer necessary for
particle_density because we are using a pre-converted CGS field as input, and
then we divide by CellVolume, which is also CGS.

--HG--
branch : yt

+++ yt/frontends/art/fields.py
@@ -280,12 +280,12 @@ add_field("particle_cell_mass_dm", function=mass_dm, units = r"\mathrm{M_{sun}}"
         projection_conversion="1")
 
 def _spdensity(field, data):
-    grid_mass = np.zeros(data.ActiveDimensions, dtype='float32')
+    grid_mass = np.zeros(data.ActiveDimensions, dtype='float64')
     if data.star_mass.shape[0] ==0 : return grid_mass 
     amr_utils.CICDeposit_3(data.star_position_x,
                            data.star_position_y,
                            data.star_position_z,
-                           data.star_mass.astype('float32'),
+                           data.star_mass,
                            data.star_mass.shape[0],
                            grid_mass, 
                            np.array(data.LeftEdge).astype(np.float64),
